requirements.txt
PsychoPy~=2021.2.3
numpy~=1.20.3
brainflow~=4.8.1
pandas~=1.3.4
pyserial~=3.5
nptyping~=1.4.4
mne~=0.24.1
scikit-learn~=1.0.1
matplotlib~=3.5.0
scipy~=1.7.3
scikit-optimize